+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [0.0.1 - 2024-06-26]
+## [0.0.1 - 2024-06-27]
 ### Added
 - First version of the project
 - Spring Application
@@ -136,4 +136,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Check export execution status
 - Coder Service
 - RStudio in Coder
-- Researcher environment project type
+- Research environment project type